Project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "authlib>=1.6.5",
    "fastapi[standard]>=0.119.1",
    "google-auth>=2.42.0",
    "huggingface-hub>=0.36.0",
    "numpy>=2.3.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=7.0.1",
    "snntorch>=0.9.4",
    "spikingjelly>=0.0.0.0.14",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "uvicorn>=0.38.0",
]
